device controller service and repository added. Client errors handling added, raised conditions in services fixed

+++ src/main/java/by/bsu/detailstorage/service/DetailService.java
@@ -5,6 +5,7 @@ import by.bsu.detailstorage.model.Detail;
 
 import jakarta.persistence.EntityExistsException;
 import jakarta.persistence.EntityNotFoundException;
+import org.hibernate.exception.ConstraintViolationException;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
@@ -12,14 +13,16 @@ import org.springframework.transaction.annotation.Transactional;
 
 import java.util.List;
 
+import static by.bsu.detailstorage.registry.EntityNameRegistry.DETAIL;
+import static by.bsu.detailstorage.registry.ErrorMessagesRegistry.ENTITY_NOT_FOUND;
+
 
 @Service
 @Transactional
 public class DetailService {
 
-    DetailRepository detailRepository;
-    private static final String ENTITY_NOT_FOUND_MSG = "Detail with id : %s not found";
-    private static final String ENTITY_EXISTS_MSG = "Detail with name : %s already exists";
+    private final DetailRepository detailRepository;
+
 
     @Autowired
     public DetailService(DetailRepository detailRepository) {
@@ -32,19 +35,19 @@ public class DetailService {
             return detailRepository.findById(id);
         }
         else {
-            throw new EntityNotFoundException(String.format(ENTITY_NOT_FOUND_MSG, id));
+            throw new EntityNotFoundException(String.format(ENTITY_NOT_FOUND.getMessage(), DETAIL.getEntityName(), id));
         }
     }
 
 
     public Detail createDetail(Detail detail) {
         detail.setName(detail.getName().trim().toLowerCase());
-        if (detailRepository.findByName(detail.getName()) == null) {
+        try {
             detailRepository.create(detail);
             return detail;
         }
-        else {
-            throw new EntityExistsException(String.format(ENTITY_EXISTS_MSG, detail.getName()));
+        catch (ConstraintViolationException e){
+            throw new EntityExistsException(String.format(ENTITY_NOT_FOUND.getMessage(), DETAIL.getEntityName(), detail.getName()));
         }
     }
 
@@ -54,14 +57,19 @@ public class DetailService {
             return detailRepository.update(detail);
         }
         else {
-            throw new EntityNotFoundException(String.format(ENTITY_NOT_FOUND_MSG, id));
+            throw new EntityNotFoundException(String.format(ENTITY_NOT_FOUND.getMessage(), DETAIL.getEntityName(), id));
         }
 
     }
 
     public void deleteDetail(long id) {
         Detail detailForDelete = detailRepository.findById(id);
-        detailRepository.delete(detailForDelete);
+        if (detailForDelete != null) {
+            detailRepository.delete(detailForDelete);
+        }
+        else {
+            throw new EntityNotFoundException(String.format(ENTITY_NOT_FOUND.getMessage(), DETAIL.getEntityName(), id));
+        }
     }
 
     public List<Detail> findMultipleDetails(Pageable pageable) {
